TriageAgent/backend/requirements.txt
# Core Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-core>=0.1.0

# LLM APIs
google-genai>=0.3.0

# NLP & NER
spacy>=3.7.0
# Run: python -m spacy download en_core_web_sm

# Firebase
firebase-admin>=6.4.0

# Utilities
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
aiofiles>=23.2.0
email-validator>=2.1.0

# Monitoring & Logging
python-json-logger>=2.0.7

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3